EOS-27546: Silence 'Defaulting container name' log collection messages
When `kubectl exec` or `kubectl cp` are used without a container name, a
warning is printed and the first container is used automatically. This
is fine but the warnings might confuse the status of the results so we
prefer to avoid them.

For support bundle generation, any of the pod's containers are a valid
selection, so use the first one listed.

Note that for kubectl v1.21+, the
`kubectl.kubernetes.io/default-container` annotation could be used to
select a default container, thus avoiding any special logic to pick a
container.

Signed-off-by: Keith Pine <keith.pine@seagate.com>