CORTX-28899: Ability to create multiple LoadBalancer IO services
Helm chart updated to allow specification of an IO service count. If the
service type is LoadBalancer, `count` instances of IO services will be
created. This supports redundancy in a cloud environment.

If the service type is `ClusterIP` or `NodePort`, the count is ignored
and only one service is created.

Signed-off-by: Keith Pine <keith.pine@seagate.com>