CORTX-29234: Default nodePorts to `null`.
Cosmetic change only, no functional difference. Since the port value is
an integer value, `null` makes a little more sense to use than an empty
string.

`""`, '~'`, `null` or the ommision of the node port values are all valid
and have the exact same behavior, which is to result in the assignment
of random port numbers.

Signed-off-by: Keith Pine <keith.pine@seagate.com>