CORTX-29122: Restore CORTX passwords
We have decided that rather than make such a significant
change as generating passwords by default that we would
first issue a written warning that this will happen in the
next release.

This changeset restores the CORTX passwords that were removed
as part of the prior CORTX-29122 change.  This change
will be reverted in the next release in a few weeks after
the QA and DevOps teams have had a chance to accommodate
these changes.

Signed-off-by: Walter Lopatka <wlopatka@gmail.com>