Fix SNS validation check
Node counting was incorrectly adding any field with the string "node" in
it (e.g. "nodePorts"). Since the SNS check depends on knowing the number
of nodes and CVGs, the calculation was wrong.

The node counting will fail in the future if any keys in addition to
"name" are added to nodes.

Signed-off-by: Keith Pine <keith.pine@seagate.com>