CORTX-31239: Support variable timeout values in deploy-cortx-cloud.sh
There is a "wait" after each of the cortx helm charts
are deployed.  This wait is guarded by a timeout.
Prior to this change the timeout values were hard-coded.

With this change in environment variable can override
the timeout value.  The supported environment
variables are:

  CORTX_DEPLOY_CONTROL_TIMEOUT (default 300s)
  CORTX_DEPLOY_DATA_TIMEOUT (default 300s)
  CORTX_DEPLOY_SERVER_TIMEOUT (default 300s)
  CORTX_DEPLOY_HA_TIMEOUT (default 120s)

This value is passed directly to "kubectl wait --timeout=${TIMEOUT}"

Signed-off-by: Walter Lopatka <wlopatka@gmail.com>