CORTX-31831: Add Kafka init container to wait for Zookeeper
Use an init container to prevent Kafka containers to start
before Zookeeper is ready for them.  This prevents Kafka
from doing a restart right out of the box.  This is does
not fix any functional problem (it is ok if the Kafka container
restarts), but it looks nicer and better controlled.

Signed-off-by: Walter Lopatka <wlopatka@gmail.com>