CORTX-32063: Parallelize support-bundle generation (#291)
Run the log gathering process for each pod in parallel.
This is needed especially for the "support_bundle generate"
calls, which can take more than 2 minutes for cortx-data pods.

The key change here is that code was refactored so that all log
collection (including support bundle generation) for a single pod
is moved to one function, and now that function is called as a
background process.  Once all pod logs/support bundles have
been collected then they are combined into a single tar file.

Signed-off-by: Walter Lopatka <wlopatka@gmail.com>