CORTX-31831: Add 60s Zookeeper connect timeout to Kafka container
Kafka depends on Zookeeper being active and ready to receive
connections.  If it cannot connect to Zookeeper within a specified
timeout then it will terminate, allowing K8s to restart Kafka.
This behavior is fine, but it will look nicer if Kafka can
wait longer before terminating so that it does not show any
restarts on deploy.

This changeset increases the Zookeeper connect timeout from the
default of 6s to 60s, which looks like plenty of time for
Zookeeper to start up in a CORTX environment.

Signed-off-by: Walter Lopatka <wlopatka@gmail.com>