CORTX-30966: Support SSL Certificate as a Secret
This changeset updates the deploy support for the SSL
certficate that is used for cortx-control (CSM) and
cortx-server (RGW) pods.

In solution.yaml, if common.ssl.secret is specified, then
a pre-built SSL certifiate will be used to create the Secret
that holds the certificate.

Alternatively, if common.ssl.externa_secret is specified, then
CORTX will use the certificate contained in the specified
secret.

The specified secret must be in the form of a pem file that
contains both the certifcate and private key.

Signed-off-by: Walter Lopatka <walter.lopatka@seagate.com>