CORTX-34284: Do not count cortx-consul-client pods in status check (#380)
status-cortx-cloud.sh was checking the total count of consul pods
against the sum of expected server pods (max(3, #worker-nodes))
and expected client pods.  However, the expected number is based
on the number of worker nodes not tainted w/ NoSchedule.  This
does not work as expected when a node that is normally a worker
node is tainted with NoSchedule.

To fix I have stopped checking for the expected number of
cortx-consul-client pods.  This is not a useful check anyway.

Signed-off-by: Walter Lopatka <walter.lopatka@seagate.com>

Signed-off-by: Walter Lopatka <walter.lopatka@seagate.com>
Co-authored-by: Rick Osowski <rosowski@gmail.com>